refact(FindProperty): inline all the things
