+++ editor.planx.uk/package.json
@@ -61,6 +61,7 @@
     "scroll-into-view-if-needed": "^2.2.26",
     "sharedb": "^1.4.2",
     "subscriptions-transport-ws": "^0.9.18",
+    "swr": "^0.4.0",
     "yup": "^0.32.8",
     "zustand": "^3.1.4"
   },
