--- editor.planx.uk/src/@planx/components/FindProperty/Public/PropertyInformation/styles.ts
@@ -1,26 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-export const constraintsStyles = makeStyles((theme) => ({
-  constraint: {
-    borderLeft: `3px solid rgba(0,0,0,0.3)`,
-    padding: theme.spacing(1, 1.5),
-    marginBottom: theme.spacing(0.5),
-  },
-}));
-export const propertyDetailStyles = makeStyles((theme) => ({
-  propertyDetail: {
-    display: "flex",
-    justifyContent: "flex-start",
-    borderBottom: `1px solid ${theme.palette.background.paper}`,
-  },
-}));
-export const propertyInformationStyles = makeStyles((theme) => ({
-  map: {
-    padding: theme.spacing(1, 0),
-  },
-  constraint: {
-    borderLeft: `3px solid rgba(0,0,0,0.3)`,
-    padding: theme.spacing(1, 1.5),
-    marginBottom: theme.spacing(0.5),
-  },
-}));
