+++ editor.planx.uk/src/@planx/components/FindProperty/model.ts
@@ -7,3 +7,19 @@ export const parseFindProperty = (
 ): FindProperty => ({
   ...parseMoreInformation(data),
 });
+
+export interface Address {
+  UPRN: number;
+  team: string;
+  sao: string;
+  pao: string;
+  organisation: string | null;
+  street: string;
+  town: string;
+  postcode: string;
+  blpu_code: string;
+  planx_description: string;
+  planx_value: string;
+  x: number;
+  y: number;
+}
